HW22

+++ api/src/main/java/com/gb/market/api/dto/ProductDto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,5 +14,5 @@ public class ProductDto {
 
     private String name;
 
-    private Float price;
+    private BigDecimal price;
 }
