+++ core/src/main/java/com/gb/market/core/repositories/specifications/ProductSpecifications.java
@@ -3,12 +3,14 @@ package com.gb.market.core.repositories.specifications;
 import com.gb.market.core.entities.Product;
 import org.springframework.data.jpa.domain.Specification;
 
+import java.math.BigDecimal;
+
 public class ProductSpecifications {
-    public static Specification<Product> priceGreaterOrEqualThan (Float minPrice) {
+    public static Specification<Product> priceGreaterOrEqualThan (BigDecimal minPrice) {
         return (root, query, criteriaBuilder) -> criteriaBuilder.greaterThanOrEqualTo(root.get("price"), minPrice);
     }
 
-    public static Specification<Product> priceLessOrEqualThan (Float maxPrice) {
+    public static Specification<Product> priceLessOrEqualThan (BigDecimal maxPrice) {
         return (root, query, criteriaBuilder) -> criteriaBuilder.lessThanOrEqualTo(root.get("price"), maxPrice);
     }
