+++ core/src/main/resources/db/migration/V1__init.sql
@@ -1,7 +1,7 @@
 CREATE TABLE products (
     id            bigserial  PRIMARY KEY,
     name          VARCHAR(255),
-    price         float,
+    price         numeric(8, 2),
     created_at    timestamp default current_timestamp,
     updated_at    timestamp default current_timestamp
 );
@@ -10,7 +10,7 @@ CREATE TABLE orders
 (
     id          bigserial primary key,
     username    VARCHAR (255) not null,
-    total_price float    not null,
+    total_price numeric(8, 2)    not null,
     created_at  timestamp default current_timestamp,
     updated_at  timestamp default current_timestamp
 );
@@ -21,7 +21,7 @@ CREATE TABLE order_items
     product_id        bigint not null references products (id),
     order_id          bigint not null references orders (id),
     quantity          int    not null,
-    price             float    not null,
+    price             numeric(8, 2)    not null,
     created_at        timestamp default current_timestamp,
     updated_at        timestamp default current_timestamp
 );
