+++ core/src/test/java/com/gb/market/core/ProductServiceTest.java
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -25,11 +26,11 @@ public class ProductServiceTest {
 
     @Test
     public void findProductsTest() {
-        Product productOne = new Product(1L, "One", 10f, null, null);
-        Product productTwo = new Product(2L, "Two", 20f, null, null);
+        Product productOne = new Product(1L, "One", BigDecimal.valueOf(10), null, null);
+        Product productTwo = new Product(2L, "Two", BigDecimal.valueOf(20), null, null);
         Mockito.doReturn(Optional.of(productOne)).when(productRepository).findById(1L);
         Product product = productService.findById(1L).get();
-        Assertions.assertEquals(10f, product.getPrice());
+        Assertions.assertEquals(BigDecimal.valueOf(10), product.getPrice());
 
         List<Product> productList = new ArrayList<>();
         productList.add(productOne);
@@ -37,7 +38,7 @@ public class ProductServiceTest {
         Mockito.doReturn(productList).when(productRepository).findAll();
         Assertions.assertEquals("Two", productService.findAllForWs().get(1).getName());
 
-        ProductDto productDto = new ProductDto(1L, "Three", 30f);
-        Assertions.assertEquals(30f, productService.update(productDto).getPrice());
+        ProductDto productDto = new ProductDto(1L, "Three", BigDecimal.valueOf(30));
+        Assertions.assertEquals(BigDecimal.valueOf(30), productService.update(productDto).getPrice());
     }
 }
